+++ tests/test_files/error_files/py_file_error_multiline_types.py
@@ -1,0 +1,12 @@
+"""This is an example python file"""
+
+
+class ExampleClass:
+    def __init__(self, *args, **kwargs):
+        """First doc str
+
+        :param list args: Args
+        :type args: list
+            but multiline
+
+        """